requirements.txt
future==0.18.2
numpy==1.18.5
torch==1.5.0
PyYAML~=5.3.1